Modulo 2 - Exercicio 14 resolvido

+++ modulo2/ex14/operation.h
@@ -1,4 +1,4 @@
 #ifndef OPERATION_H
 #define OPERATION_H
-	double operation();
+	int operation();
 #endif
